Added support code to set real-time priority.
The only minimum timing in the C2 protocol is for the low pulse of C2CK.
For now, we don't do anything to ensure that we don't violate that timing.
This may explain why communication sometimes fails. Using a real-time
priority should reduce the number of "surprises". But surprisingly, it
makes things worse. Needs more investigation.